grid: Rename alternate_protocols_cache.{h,cc} to alternate_protocols_cache_impl.{h,cc}
