+++ source/common/http/BUILD
@@ -161,8 +161,8 @@ envoy_cc_library(
 
 envoy_cc_library(
     name = "alternate_protocols_cache",
-    srcs = ["alternate_protocols_cache.cc"],
-    hdrs = ["alternate_protocols_cache.h"],
+    srcs = ["alternate_protocols_cache_impl.cc"],
+    hdrs = ["alternate_protocols_cache_impl.h"],
     deps = [
         "//include/envoy/common:time_interface",
     ],
