+++ source/common/http/conn_pool_grid.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common/http/alternate_protocols_cache.h"
+#include "common/http/alternate_protocols_cache_impl.h"
 #include "common/http/conn_pool_base.h"
 #include "common/http/http3_status_tracker.h"
